Moved coinlist to data storage api methodolgy
Closes #919
